requirements.txt
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
pytest==9.0.3
